Basic DAL (#4)
* Context.

* appsettings.json added

* Startup configuration for DAL.

* Seeder.

* Initial migration & fixed user's level type.

+++ RequirementsLab.Core/Entities/User.cs
@@ -5,7 +5,7 @@ namespace RequirementsLab.Core.Entities
 {
     public class User : IdentityUser<int>, IEntity
     {
-        public decimal Level { get; set; }
+        public float Level { get; set; }
 
         public string Name { get; set; }
